% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_matrix}
\alias{A_matrix}
\title{Age mixing matrix based on Milwid et al, 2022 and the Engage Cohort Study age distribution (2022)}
\format{
The list has 3 sublists (for Montreal, Toronto, and Vancouver), each contains a 5*5 matrix.
}
\usage{
A_matrix
}
\description{
Probability of sexual partnership formation between gay, bisexual, and other men who have sex with men belonging to different age groups.
}
\keyword{datasets}
