% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{case_data}
\alias{case_data}
\title{Daily confirmed mpox cases in Montréal, Toronto, and Vancouver (May-Oct 2022.}
\format{
The data frame has 513 rows and the following 9 columns:
\describe{
\item{prov}{name of the provinces: Quebec, Ontario, and British Columbia, where Montreal, Toronto, and Vancouver are located, respectively}
\item{date}{date}
\item{time_conti}{day since first reported case in each city}
\item{prov_incidence}{provincial daily mpox incidence}
\item{prov_cumul_cases}{provincial cumulative mpox incidence}
\item{city_prov}{city fraction of provincial mpox incidence, varying weekly}
\item{confirmed_total}{confirmed fraction of confirmed + probable mpox incidence}
\item{first_dose}{mpox vaccine doses administered, varying weekly and allocated uniformly to each day}
}
}
\usage{
case_data
}
\description{
case data derived from the Public Health Agency of Canada and city surveillance reports
}
\keyword{datasets}
