% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_rate_prop}
\alias{contact_rate_prop}
\title{Aggregate-level socio-demographic and behavioural data derived from the Engage Cohort Study ( 2017 - 2023 )}
\format{
The data frame has 450 rows and the following 10 columns:
\describe{
\item{city}{name of the city, Montreal, Toronto, or Vancouver}
\item{age_cats}{age groups}
\item{hiv_cats}{HIV-serostatus, 0: seronegative/unknown, 1: seropositive}
\item{sa_cats}{sexual acitivty groups acertained by on a type (ttl/anal) of sexual partners}
\item{mean_rate}{contact rate per past 6 months before the mpox outbreak for age group a, sexual activity group s, and HIV-serostatus h}
\item{prop}{proportion of the size of an age-sexual activity-HIV-serostatus group relative to the MSM population size in a city for a type (ttl or anal) of sexual partners}
\item{min_grp}{minimum number of sexual partner numbers in the past 6 months of a sexual activity group}
\item{max_grp}{maximum number of sexual partner numbers in the past 6 months of a sexual activity group}
\item{type}{ttl: all-type sexual partners, anal: anal sexual partners}
\item{c_ash}{contact rate per day before the mpox outbreak for age group a, sexual activity group s, and HIV-serostatus h}
}
}
\usage{
contact_rate_prop
}
\description{
Engage Cohort Study is a prospective,
population-based cohort study of GBM in Montréal, Toronto, and Vancouver.
Eligible participants were self-identified cis or trans men living in one of the three cities,
aged ≥16 years, who reported sex with another man in the past 6 months (P6M),
understood English or French, and provided written consent. From February 2017 to August 2019,
participants were recruited using respondent-driven sampling (RDS),
a method used to sample hard-to-reach populations to estimate more representative population characteristics.
Initial participants were purposively selected to represent diverse characteristics of the GBM community,
and all participants were invited to recruit up to six peers in their social networks.
}
\keyword{datasets}
