% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_load_fns.R
\name{.onAttach}
\alias{.onAttach}
\title{Functions called when package is attached}
\usage{
.onAttach(libname = find.package("MpoxModelPack"), pkgname = "MpoxModelPack")
}
\arguments{
\item{libname}{Defaults is find.package( "MpoxModelPack")}

\item{pkgname}{Default is "MpoxModelPack"}
}
\value{
None
}
\description{
Calls to functions to be loaded on package attachment.
}
