% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_fn_seir_model.R
\name{fn_model}
\alias{fn_model}
\title{fn_model}
\usage{
fn_model(
  city,
  import_cases_city,
  bbeta_city,
  omega_city,
  RR_H_city,
  RR_L_city,
  gamma1_city,
  period_city = 150,
  TRACING = 1,
  VACCINATING = 1,
  cpp
)
}
\arguments{
\item{city}{model one city at a time: "mtl" = Montreal, "trt" = Toronto, "van" = Vancouver}

\item{import_cases_city}{number of imported cases}

\item{bbeta_city}{transmission probability per effective contact}

\item{omega_city}{mixing parameters}

\item{RR_H_city}{rate ratio of change in sexual partner numbers among GBM with > 7 sexual partners before 2022}

\item{RR_L_city}{rate ratio of change in sexual partner numbers among GBM with <= 7 sexual partners before 2022}

\item{gamma1_city}{1/duration of effective infectious period}

\item{period_city}{days from the first reported cases to the last modeled case}

\item{TRACING}{1: contact tracing and isolating GBM; 0: no contact tracing and isolation}

\item{VACCINATING}{1: vaccinating GBM with 1-dose Imvamune vaccines; 0: no vaccination}

\item{cpp}{1: use Rcpp; 0: use R}
}
\value{
output_t a list of cases (daily reported cases across time), time (days since first reported case), X (array of size in each compartment across time), inc (daily incidence across time)
}
\description{
Loads the model function
}
\author{
: Fanyu Xiu, Carla Doyle
}
