% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fn_calibration.R
\name{getci}
\alias{getci}
\title{getci}
\usage{
getci(df)
}
\arguments{
\item{df}{a dataframe object}
}
\value{
a credible interval (vector) of the estimates
}
\description{
Function for obtaining 95\% credible interval of the estimates
}
