% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fn_set_population.R
\name{init.pop.fn}
\alias{init.pop.fn}
\title{init.pop.fn}
\usage{
init.pop.fn(city = c("mtl", "trt", "van"), type = 1)
}
\arguments{
\item{city}{"mtl" = Montreal, "trt" = Toronto, "van" = Vancouver (can input a vector), Default: c("mtl", "trt", "van")}

\item{type}{1 = all-type, 2 = anal, Default: 1}
}
\value{
pop_dat
}
\description{
Sets up the initial model population of interest (according to city and type of sexual partners).
}
\examples{
\dontrun{
if(interactive()){
 init.pop.fn(city = c("trt", "van"), type = 1)
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
\author{
: Fanyu Xiu, Carla Doyle
}
