% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_fn_calibration_sen.R
\name{llk_all_sen}
\alias{llk_all_sen}
\title{llk_all_sen}
\usage{
llk_all_sen(
  theta,
  cal_type = 1,
  cal_cpp,
  VE_llk,
  contact_prop_llk,
  contact_dur_llk
)
}
\arguments{
\item{theta}{value of a vector of parameter in logit scale}

\item{cal_type}{which type of sexual act to be selected (all-type:1 or anal:2), Default: 1}

\item{cal_cpp}{whether to run calibration with rcpp (1) or R (0)}

\item{VE_llk}{vaccine effectiveness to run the calibration (for sensitivity analysis)}

\item{contact_prop_llk}{proportion contact traced and isolated to run the calibration (for sensitivity analysis)}

\item{contact_dur_llk}{days it takes for contact tracing (for sensitivity analysis)}
}
\value{
a value equals to the log posterior density
}
\description{
function to output posterior density of a parameter set using negative binomial likelihood (not calibrating RR)
}
