% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_fn_load_params.R
\name{load.params.fn}
\alias{load.params.fn}
\title{load.params.fn}
\usage{
load.params.fn(
  VE = 0.5149,
  contact_prop = 0.2,
  contact_dur = 2,
  standardized_vaccine_date = F
)
}
\arguments{
\item{VE}{vaccine effectiveness from meta-analysis, Default: "0.5149"}

\item{contact_prop}{proportion of contact traced and isolated after first reported case, Default: 0.20}

\item{contact_dur}{days it takes for contact tracing, Default: 2}

\item{standardized_vaccine_date}{whether or not standardizing vaccination start dates across the cities (for sensitivity analysis)}
}
\value{
parameters for the model will be added to the global environment
}
\description{
Loads the model's fixed parameter values
}
\details{
age and HIV mixing matrices from Milwid et al, 2022 (https://bmcinfectdis.biomedcentral.com/articles/10.1186/s12879-022-07207-7#Sec18)
}
\examples{
\dontrun{
if(interactive()){
 load.params.fn(); H_matrix
 }
}
}
\seealso{
\code{\link[dplyr]{select}}
}
\author{
: Fanyu Xiu, Carla Doyle
}
