% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MpoxModelPack-package.R
\docType{package}
\name{MpoxModelPack-package}
\alias{MpoxModelPack}
\alias{MpoxModelPack-package}
\title{MpoxModelPack: Compartmental Model for Mpox Transmission among Men Who Have Sex with Men in Canada}
\description{
The goal of MpoxModelPack is to build a dynamic compartmental model that retrospectively assesses the impact of change in partner numbers, contact tracing, and vaccination, among gay, bisexual, and other men who have sex with men (GBM) during the 2022-2023 mpox outbreaks in major Canadian cities.
}
\author{
\strong{Maintainer}: Fanyu Xiu \email{vanessa.xiu@mail.mcgill.ca}

Other contributors:
\itemize{
  \item Carla Doyle \email{carla.doyle@mail.mcgill.ca} [contributor]
  \item Mathieu Maheu-Giroux \email{mathieu.maheu-giroux@mcgill.ca} [contributor]
  \item Jorge Luis Flores Anato \email{jorge.l.flores@mail.mcgill.ca} [contributor]
}

}
\keyword{internal}
