% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fn_calibration.R
\name{prior_dens_fix}
\alias{prior_dens_fix}
\title{prior_dens_fix}
\usage{
prior_dens_fix(theta)
}
\arguments{
\item{theta}{value of a vector of parameter in logit scale}
}
\value{
a value equals to log prior density
}
\description{
sum of log prior density for a given set of parameter values
}
\author{
: Mathieu Maheu-Giroux, Jorge Luis Flores Anato, Fanyu Xiu
}
