% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_fn_calibration_sen.R
\name{prior_dens_fix_sen}
\alias{prior_dens_fix_sen}
\title{prior_dens_fix_sen}
\usage{
prior_dens_fix_sen(theta)
}
\arguments{
\item{theta}{value of a vector of parameter in logit scale}
}
\value{
a value equals to log prior density
}
\description{
sum of log prior density for a given set of parameter values (not calibrating RR)
}
\author{
: Mathieu Maheu-Giroux, Jorge Luis Flores Anato, Fanyu Xiu
}
