% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fn_calibration.R
\name{simul_fun_all}
\alias{simul_fun_all}
\title{simul_fun_all}
\usage{
simul_fun_all(
  cal_type,
  hessian,
  thetas,
  sim = 1000,
  SIR = TRUE,
  nsir = 5000,
  with_replacement = TRUE,
  track_sims = FALSE,
  parallel = TRUE,
  cal_cpp,
  VE_sim,
  contact_prop_sim,
  contact_dur_sim,
  standardized_vaccine_date_sim
)
}
\arguments{
\item{cal_type}{which type of sexual act to be selected (all-type:1 or anal:2), Default: 1}

\item{hessian}{hessian matrix returned by the optim function}

\item{thetas}{maximum a posterior (mode) of parameters to be calibrated}

\item{sim}{number of resamples, Default: 1000}

\item{SIR}{use sampling importance resampling algorithm, Default: TRUE}

\item{nsir}{number of draws from the importance function, Default: 5000}

\item{with_replacement}{resample with replacement, Default: TRUE}

\item{track_sims}{tracking simulations , Default: FALSE}

\item{parallel}{to fasten the loops, Default: TRUE}

\item{cal_cpp}{whether to run calibration with rcpp (1) or R (0)}

\item{VE_sim}{vaccine effectiveness to run the calibration (for sensitivity analysis)}

\item{contact_prop_sim}{proportion contact traced and isolated to run the calibration (for sensitivity analysis)}

\item{contact_dur_sim}{days it takes for contact tracing to run the calibration (for sensitivity analysis)}

\item{standardized_vaccine_date_sim}{whether or not standardizing vaccination start dates across the cities (for sensitivity analysis)}
}
\value{
a list of result (model fit for each parameter sets), par_ci (posterior CrI for parameters), AF_ci (posterior CrI of averted fractions), samples (all resampled parameter sets)
}
\description{
simulation to obtain credible interval of all parameter sets (CrI)
}
\seealso{
\code{\link[Matrix]{solve-methods}}, \code{\link[Matrix]{nearPD}}
\code{\link[mvtnorm]{Mvnorm}}, \code{\link[mvtnorm]{Mvt}}
\code{\link[parallel]{makeCluster}}, \code{\link[parallel]{detectCores}}, \code{\link[parallel]{clusterApply}}
\code{\link[doParallel]{registerDoParallel}}
\code{\link[Rcpp]{sourceCpp}}
\code{\link[foreach]{foreach}}
}
